commit changes

+++ app/src/main/java/com/example/leedstrinity/hospimanagementapp/data/entities/Patient.java
@@ -65,7 +65,7 @@ public class Patient {
 
     // --- Getters & Setters ---
     public long getId() { return id; }
-    public void setId(int id) { this.id = id; }
+    public void setId(long id) { this.id = id; }
 
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
